% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.R
\name{myttest}
\alias{myttest}
\title{Two-sample t-test function}
\usage{
myttest(x, y, alpha = 0.05, paired = FALSE)
}
\arguments{
\item{x}{A numeric vector of data for group 1}

\item{y}{A numeric vector of data for group 2}

\item{alpha}{A numeric value specifying the level of significance (default is 0.05)}

\item{paired}{A logical indicating whether the samples are paired (default is FALSE)}
}
\value{
A list containing the results of the t-test, including the test type, summary statistics, null hypothesis rejection, and input data
}
\description{
Performs a two-sample t-test on x and y and returns the results as an object of class "Rttest"
}
\examples{
# X Variable
set.seed(21)
x <- rnorm(30,5,2)

# Y Variable
set.seed(23)
y <- rnorm(30, 3,2)

# Alpha level
alpha <- 0.05

# Call the contructor
obj <- myttest(x = x, y = y, alpha = 0.05)

# Print object
print(obj)
}
