% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Plot Method for Rttest Objects}
\usage{
plot(Rttest_obj)
}
\arguments{
\item{Rttest_obj}{An object of class Rttest.}
}
\value{
The result of the specific Plot method for the class of the input object.
}
\description{
This function creates a generic method for Ploting Rttest objects. Uses ggplot2
}
