% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Rttest.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{Plot Method for Rttest Objects}
\usage{
\method{plot}{Rttest}(Rttest_obj)
}
\arguments{
\item{Rttest_obj}{An object of class Rttest.}
}
\value{
The result of the specific Plot method for the class of the input object.
}
\description{
This function creates a generic method for Ploting Rttest objects. Uses ggplot2
}
\examples{
# Create a sample Rttest object
# X Variable
set.seed(21)
x <- rnorm(30,5,2)

# Y Variable
set.seed(23)
y <- rnorm(30, 3,2)

# Alpha level
alpha <- 0.05

# Call the contructor
obj <- myttest(x = x, y = y, alpha = 0.05)

# Plot the sample Rttest
plot(obj)
}
\seealso{
\code{\link{myttest}}

\code{\link{plot}}
}
