% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Rttest.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Print Rttest Objects in a Readable Format}
\usage{
\method{print}{Rttest}(Rttest_obj, roundToDigits = 3)
}
\arguments{
\item{Rttest_obj}{An object of class Rttest.}

\item{roundToDigits}{Number of digits to round the output to.}
}
\value{
A formatted table displaying the data from the Rttest object.
}
\description{
This function defines the print method for Rttest objects. It extracts the data from the object and
prints it in a readable format using the kable function from the knitr package.
}
\examples{
# Create a sample Rttest object
sample_data <- data.frame(a = 1:5, b = 6:10)
sample_Rttest_obj <- list(data = sample_data, class = "Rttest")

# Print the sample Rttest object
print(sample_Rttest_obj)
}
