% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ggplot.R
\name{theme_ggplot}
\alias{theme_ggplot}
\title{Function that returns a theme object for ggplot objects}
\usage{
theme_ggplot()
}
\value{
a theme via a list of geoms that can be added to ggplot2 objects
}
\description{
Function that returns a theme object for ggplot objects
}
\examples{

library(ggplot2)

mtcars |>
  ggplot(aes(x=wt, y=mpg)) + geom_point() +
  theme_ggplot()
}
