% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\title{Print Method for Rttest Objects with Optional Rounding}
\usage{
print(Rttest_obj, roundToDigits = 3)
}
\arguments{
\item{Rttest_obj}{An object of class Rttest.}

\item{roundToDigits}{The number of decimal places to round the data values to (default: 3).}
}
\value{
The result of the specific print method for the class of the input object.
}
\description{
This function creates a generic method for printing Rttest objects. It allows for optional rounding
of the data values by specifying the number of digits to round to.
}
\seealso{
\code{\link{print.Rttest}}
}
